# docs/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: hajaCheck API
  version: "0.38.0-draft"
  description: >
    Contract-First 초안 (PRD §6, §8). 실제 소스 오브 트루스는 이 파일.


    ## 실시간(WS) 상담 — /ws STOMP (FR-7, #20/HAJA-33, #1002/HAJA-496)

    OpenAPI 3.0 은 WebSocket/STOMP 를 네이티브로 표현하지 못하므로, 이 계약은 `paths` 항목이 아니라
    이 서술로 대체한다(`backend/.../counsel/websocket/WebSocketConfig.java` 실구현 기준).

    - **엔드포인트**: `/ws` (SockJS 미사용 — dev(Vite)/prod(nginx) 동일 오리진 프록시 전제, 기본 same-origin
      정책 유지).
    - **인증**: SESSION 쿠키 기반, 2단계로 강제한다 — ① 핸드셰이크(`CounselHandshakeInterceptor`)에서
      SESSION 쿠키 검증, ② 인바운드 채널(`StompAuthChannelInterceptor`)에서 CONNECT 재검증 + SUBSCRIBE
      시점 당사자(참여자) 검증. 클라이언트 페이로드의 senderId 는 신뢰하지 않고 핸드셰이크에서 확립한
      `StompUserPrincipal` 로만 발신자를 식별한다.
    - **destinations**:
      - `SEND /app/counsel/{ticketId}/send` — 상담 메시지 전송(`ChatMessageSendRequest`: content
        ≤2000자/attachmentKey ≤500자, 텍스트·첨부 중 최소 하나). 티켓이 IN_PROGRESS 가 아니거나 발신자가
        참여자(사용자 본인/담당 상담원)가 아니면 예외 대신 조용히 드롭한다(WS 프레임 처리 중 예외로 세션이
        끊기는 것을 피하기 위한 방어적 설계).
      - `SUBSCRIBE /topic/counsel/{ticketId}` — 해당 상담방 브로드캐스트 구독. 서버가 저장된 메시지를
        `ChatMessageResponse` 로 이 destination 에 `convertAndSend` 한다.
      - `SUBSCRIBE /user/queue/counsel/assigned` — 개인 알림(user destination prefix `/user`). 내 티켓이
        상담원에게 배정되면(`POST /api/counsel/tickets/{id}/assign` 성공 시) `CounselTicketResponse` 를
        수신한다.
      - `SUBSCRIBE /user/queue/counsel/ended` — 개인 알림. 내 티켓이 종료되면(`POST /api/counsel/tickets/
...


...


...


...


...


...


...


...


...

...

...

...
servers:
...
paths:
...
  /api/me/plan:
    get:
      summary: 내 플랜 + 사용량 조회 (마이페이지, HAJA-177)
      description: >
        구독 소유 판별은 companyId 있으면 회사 구독(user_plans.company_id), 없으면 개인 구독(user_plans.user_id).
        "현재 구독"은 ACTIVE 우선, 없으면 UPGRADE_REQUESTED 순으로 조회하고 둘 다 없으면(EXPIRED 만 존재)
        PLAN_NOT_FOUND(404)다. limits.* 는 plans.max_* 가 null(무제한)이면 null 을 그대로 반환한다.
        `MembershipController.getMyPlan` 구현됨.
      security:
        - SessionCookie: []
      tags: [membership]
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyPlanApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 활성 구독 없음 (PLAN_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/me/seats:
    get:
      summary: 좌석 현황 조회 (마이페이지, HAJA-177)
      description: >
        개인 사용자(companyId=null)는 본인 1인만 담은 좌석 목록을 반환한다. 회사 소속 사용자는 회사 소속
        ACTIVE 사용자 목록(최대 200건, MEMBERSHIP_SEATS_MAX)을 반환하되 used(총원 수)는 이 상한과 무관하게
        정확한 count 쿼리로 산출한다. 이메일은 비마스킹(동일 회사 소속만 조회 가능한 내부 명부이므로).
        `MembershipController.getSeats` 구현됨.
      security:
        - SessionCookie: []
      tags: [membership]
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeatsApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 활성 구독 없음 (PLAN_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
...
  /api/me/plan/upgrade-inquiry:
    post:
      summary: 업그레이드 문의 (마이페이지, HAJA-177)
      description: >
        PG 실결제 대체 — 구독 상태를 UPGRADE_REQUESTED 로 전이한다(멱등). 회사 구독은 회사 대표
        (company.ownerUserId)만, 개인 구독은 그 구독의 소유 사용자만 요청할 수 있다(PLAN_FORBIDDEN).
        `MembershipController.requestUpgrade` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [membership]
      responses:
        "200":
          description: 문의 접수 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpgradeInquiryApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: 구독 소유자 아님 (PLAN_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 활성 구독 없음 (PLAN_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/admin/plans:
    get:
      summary: 요금제 카탈로그 조회 (관리자 플랜·쿼터, FR-8-A #507)
      description: >
        제공 중인 전체 요금제 목록(플랜 변경 선택지). ADMIN 전용 — `/api/admin/**` 는 SecurityConfig 필터에서
        hasRole(ADMIN) 로 강제한다. `AdminPlanController.getPlanCatalog` 구현됨.
      security:
        - SessionCookie: []
      tags: [admin]
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPlanCatalogApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: ADMIN 아님 (FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
...
  /api/admin/plan-quota:
    get:
      summary: 회사 멤버별 쿼터 사용 현황 조회 (관리자 플랜·쿼터, FR-8-A #507, frontend #508)
      description: >
        요청 관리자 회사에 소속된 활성(status=ACTIVE) 멤버별 이번 달 쿼터 사용량과 KPI 통계를 반환한다.
        plan/quotaLimit 은 회사가 구독한 단일 플랜값이라 모든 행에서 동일하고(company_memberships 상속
        아님 — user_plans 는 회사 단위 1건), 활성 구독이 없으면(PLAN_NOT_FOUND 상황) 목록 조회 자체는
        실패시키지 않고 plan/quotaLimit/companyPlan 을 전부 null 로 반환한다 — `/api/admin/plan`·
        `/api/admin/plan/history` 와 달리 "구독 없음"도 정상 상태로 보여줘야 하는 목록 뷰이기 때문이다.
        quotaUsed 는 담당 점검자(inspections.assigned_inspector_id) 단위 media 집계 근사치(멤버별 실제
        업로더 FK 부재 — MediaRepository 문서 참고). stats(activeUsers/totalQuotaUsagePercent/companyPlan)는
        검색어(keyword)와 무관하게 회사 전체 기준으로 집계한다. page 는 1부터 시작.
        `AdminPlanController.getPlanQuota` 구현됨.
      security:
        - SessionCookie: []
      tags: [admin]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 10, minimum: 1 }
        - name: keyword
          in: query
          description: 멤버 이름·이메일 부분일치 검색(대소문자 무시)
          schema: { type: string }
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPlanQuotaApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: ADMIN 아님 / 회사 없는 관리자 (FORBIDDEN / PLAN_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 요청 관리자 본인의 회사 멤버십이 유효 승인 상태가 아님 (PLAN_NOT_FOUND) — 회사 자체의 구독 유무와는 별개
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/platform-admin/plans-quota:
    get:
      summary: 전사 사용자별 플랜·쿼터 사용 현황 조회 (플랫폼 관리자, #624, frontend #625)
      description: >
        companyId 스코프 없이 전사 사용자(PLATFORM_ADMIN 자신 제외)를 사용자별로 나열한다. companyId/
        companyName 은 소속 회사(회사 미소속이면 둘 다 null), plan/quotaLimit/remainingDays 는 그 사용자가
        소속된 회사의 활성(ACTIVE) 구독에서 나온다 — 회사마다 구독 플랜이 다를 수 있어(단일 회사 스코프인
        /api/admin/plan-quota 와 달리) 행마다 값이 다를 수 있다. 회사 미소속(개인 계정)이거나 소속 회사에
        활성 구독이 없으면 plan/quotaLimit/remainingDays 전부 null 이고 status 는 EXPIRED. remainingDays 는
        `user_plans.current_period_end`(#1104/HAJA-525 로 실체화된 컬럼) 기준 잔여 일수다 — 과거의
        `startedAt + 1개월` 조회 시점 파생 근사 계산은 폐기됐다.
        ⚠️ **remainingDays 가 null 이라고 만료로 판정하지 말 것** — null 은 "만료"와 "무기한(FREE, 종료일
        자체가 없음)" 두 경우를 모두 뜻한다. FREE 구독은 `remainingDays=null` + `status=ACTIVE` 조합으로
        내려오므로, **만료 판정은 반드시 status 로** 한다. status 는 만료면 EXPIRED, 갱신 임박(남은 7일
        이하)이면 WARNING, 그 외 ACTIVE. stats(activeUsers/totalQuotaUsagePercent)는 검색어(keyword)·플랜(plan) 필터와 무관하게 전사
        기준으로 집계하며, totalQuotaUsagePercent 는 회사마다 한도가 다를 수 있어 유효 한도를 가진 사용자별
        사용률의 평균이다. page 는 1부터 시작. `PlatformAdminPlanQuotaController.getPlanQuota` 구현됨.
      security:
        - SessionCookie: []
      tags: [platform-admin]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 10, minimum: 1 }
        - name: keyword
          in: query
          description: 사용자 이름·이메일·소속 기업명 부분일치 검색(대소문자 무시)
          schema: { type: string }
        - name: plan
          in: query
          description: 소속 회사의 구독 플랜으로 좁혀본다(회사 단위 필터 — 해당 플랜을 구독 중인 회사의 사용자만 반환).
          schema: { type: string, enum: [FREE, STANDARD, ENTERPRISE] }
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformAdminPlanQuotaApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: PLATFORM_ADMIN 아님 (FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/platform-admin/plans:
    get:
      summary: 요금제 정책 카탈로그 조회 (플랫폼 관리자, #624 후속)
      description: >
        FREE/STANDARD/ENTERPRISE 전체 요금제의 가격·한도·기능 제공 여부를 반환한다(PLATFORM_ADMIN 전용).
        `AdminPlanController`(#507, "/api/admin/plans")와 같은 plans 테이블을 보되 편집 권한만 다르다.
        `PlatformAdminPlanPolicyController.getCatalog` 구현됨.
      security:
        - SessionCookie: []
      tags: [platform-admin]
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformAdminPlanCatalogApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: PLATFORM_ADMIN 아님 (FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
    put:
      summary: 요금제 정책 일괄 변경 (플랫폼 관리자, #624 후속, 사용자 지시)
      description: >
        FREE/STANDARD/ENTERPRISE 3개 플랜의 가격·한도·기능 제공 여부를 한 번에 원자적으로 교체한다
        (PLATFORM_ADMIN 전용). plans 배열은 3개 플랜명을 각각 정확히 한 번씩 포함해야 한다(누락·중복·
        미지원 플랜명은 400 PLAN_POLICY_INVALID). hasAiAddon은 이 화면의 편집 대상이 아니라 요청에
        없으며, 변경 후에도 기존 값이 그대로 유지된다. plans 는 회사 스코프 없는 전역 참조 데이터라
        변경 즉시 그 플랜을 구독 중인 모든 회사/개인의 한도·가격 표시에 반영된다(user_plans 자체는
        건드리지 않음 — 플랜 변경을 일으키지 않는다). `PlatformAdminPlanPolicyController.updatePolicies` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [platform-admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformAdminPlanPolicyUpdateRequest" }
      responses:
        "200":
          description: 변경 성공 — 갱신된 카탈로그 반환
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformAdminPlanCatalogApiResponse" }
        "400":
          description: 입력값 오류(INVALID_INPUT) 또는 3개 플랜 구성 오류(PLAN_POLICY_INVALID)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: PLATFORM_ADMIN 아님 (FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
...
components:
  securitySchemes:
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Spring Boot가 FastAPI 내부 호출에 부착하는 AI_INTERNAL_KEY
    InternalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
      description: Spring Boot와 FastAPI 환경변수에만 주입하는 내부 호출 토큰
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Security 인증 세션 쿠키
    XsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: 상태 변경 요청에 필요한 double-submit CSRF 토큰
  schemas:
...
    ApiError:
      type: object
      properties:
        code: { type: string, example: AUTH_INVALID_CREDENTIALS }
        message: { type: string, example: 아이디 또는 비밀번호가 올바르지 않습니다. }
...
    ErrorApiResponse:
      type: object
      description: 실패 envelope — error=ApiError
      properties:
        success: { type: boolean, example: false }
        data: { type: object, nullable: true, example: null }
        error: { $ref: "#/components/schemas/ApiError" }

    # ── counsel 도메인 (FR-7, #20/HAJA-33, #1002/HAJA-496) ──
...
    MyPlanResponse:
      type: object
      required: [plan, limits, usage]
      properties:
        plan:
          type: object
          required: [name, priceMonthly, status, nextBillingDate, businessVerified]
          properties:
            name: { type: string, enum: [FREE, STANDARD, ENTERPRISE], example: STANDARD }
            priceMonthly: { type: number, example: 99000 }
            status: { type: string, enum: [ACTIVE, EXPIRED, UPGRADE_REQUESTED], example: ACTIVE }
            paymentPendingUntil:
              type: string
              format: date-time
              nullable: true
              description: >
                미결제 유예 마감(#1177). NULL = 정상. 값이 있으면 **유료→유료 하향 예약이 적용됐으나 아직
                정산 결제가 되지 않은 상태**이며, 그 시각까지 결제하지 않으면 FREE 로 강등된다.


                ⚠️ 유예 중에는 `limits.*` 와 요금제 기능 플래그(상담사 연결·AI 부가기능)가 **FREE 기준**으로
                내려간다 — `name` 은 구독 요금제(예: STANDARD)를 그대로 보여주므로, 이 필드가 없으면 화면은
                "STANDARD 인데 좌석 한도가 1"이라는 모순을 설명할 수 없다. 유예 진입 시점에 초과 좌석이
                FREE 기준으로 정지되며, 결제해도 정지된 구성원은 자동 복구되지 않는다(재활성화 필요).
              example: "2026-08-05T00:00:00Z"
            nextBillingDate:
              type: string
              format: date
              nullable: true
              description: >
                `user_plans.current_period_end`(#1104/HAJA-525 로 실체화된 컬럼)를 KST 날짜로 노출한다.
                FREE 플랜이면 null(무기한, #711). 과거처럼 `startedAt + 1개월`을 조회 시점에 파생 계산하지
                않는다 — 결제 승인 전이 시 리셋(now~now+1개월)되고 관리자 플랜 변경(무결제) 시에는 기존
                주기가 승계되므로, 유료여도 startedAt 과 무관한 값일 수 있다.
              example: "2026-08-15"
            businessVerified:
              type: boolean
              nullable: true
              description: >
                회사 구독이면 Company.verificationStatus == VERIFIED 여부. 개인 구독(companyId=null)이면
                사업자 인증 개념이 없으므로 null(#711).
              example: true
        limits:
          type: object
          description: plans.max_* 가 null(무제한)이면 null 을 그대로 반환한다.
          required: [maxFacilities, maxMonthlyAnalyses, maxSeats]
          properties:
            maxFacilities: { type: integer, nullable: true, example: 20 }
            maxMonthlyAnalyses: { type: integer, nullable: true, example: 100 }
            maxSeats: { type: integer, nullable: true, example: 5 }
        usage:
          type: object
          required: [facilityCount, analyzedImageCount, seatCount, period]
          properties:
            facilityCount: { type: integer, example: 3 }
            analyzedImageCount: { type: integer, example: 12 }
            seatCount: { type: integer, example: 2 }
            period: { type: string, format: date, description: "집계 기준월 1일(KST)", example: "2026-07-01" }
...
    MyPlanApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/MyPlanResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    SeatsResponse:
      type: object
      description: >
        개인 사용자는 본인 1인만 담은 목록(used=1)을 반환한다. 이메일은 회사 내부 팀 명부이므로
        비마스킹(동일 회사 소속만 조회 가능).
      required: [used, limit, members]
      properties:
        used: { type: integer, description: "회사 소속 ACTIVE 사용자 총원 수(목록 상한과 무관한 정확한 count)", example: 3 }
        limit: { type: integer, nullable: true, example: 5 }
        members:
          type: array
          maxItems: 200
          items:
            type: object
            required: [userId, name, email, role, status]
            properties:
              userId: { type: integer, format: int64, example: 1 }
              name: { type: string, example: 김민수 }
              email: { type: string, example: member@haja.com }
              role: { type: string, enum: [ADMIN, INSPECTOR, USER, COUNSELOR, PLATFORM_ADMIN], example: USER }
              status: { type: string, enum: [ACTIVE, SUSPENDED, WITHDRAWN], example: ACTIVE }
    SeatsApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/SeatsResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    UpgradeInquiryResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ACTIVE, EXPIRED, UPGRADE_REQUESTED], example: UPGRADE_REQUESTED }
    UpgradeInquiryApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: "#/components/schemas/UpgradeInquiryResponse" }
        error: { type: object, nullable: true, example: null }

    # ── 관리자 플랜·쿼터 관리 (구현됨, backend/admin, FR-8-A #507 / HAJA-258) ──
    AdminPlanItem:
      type: object
      description: 요금제 카탈로그 항목. max_* 는 무제한이면 null.
      required:
        [id, name, maxFacilities, maxMonthlyAnalyses, maxSeats, hasPdfWatermark, hasCounselorAccess, hasAiAddon, priceMonthly]
      properties:
        id: { type: integer, format: int64, example: 2 }
        name: { type: string, enum: [FREE, STANDARD, ENTERPRISE], example: STANDARD }
        maxFacilities: { type: integer, nullable: true, example: 10 }
        maxMonthlyAnalyses: { type: integer, nullable: true, example: 300 }
        maxSeats: { type: integer, nullable: true, example: 5 }
        hasPdfWatermark: { type: boolean, example: false }
        hasCounselorAccess: { type: boolean, example: true }
        hasAiAddon: { type: boolean, example: true }
        priceMonthly: { type: number, nullable: true, example: 29000 }
    AdminPlanCatalogResponse:
      type: object
      required: [plans]
      properties:
        plans:
          type: array
          items: { $ref: "#/components/schemas/AdminPlanItem" }
    AdminPlanCatalogApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/AdminPlanCatalogResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
...
    AdminPlanQuotaMember:
      type: object
      description: 회사 소속 멤버(개인 계정) 1명의 이번 달 쿼터 사용 현황 한 행. plan/quotaLimit 은 회사 단위
        구독값이라 모든 행에서 동일(활성 구독 없으면 둘 다 null). quotaUsed 만 멤버별로 다르다(근사치).
      required: [id, name, email, plan, quotaUsed, quotaLimit]
      properties:
        id: { type: integer, format: int64, example: 12 }
        name: { type: string, example: "김민준" }
        email: { type: string, example: "minjun.kim@company.com" }
        plan: { type: string, enum: [FREE, STANDARD, ENTERPRISE], nullable: true, example: STANDARD }
        quotaUsed: { type: integer, example: 145 }
        quotaLimit: { type: integer, nullable: true, example: 1000 }
    AdminPlanQuotaStats:
      type: object
      description: KPI 카드·"현재 플랜" 카드 값 — 검색어(keyword)와 무관하게 회사 전체 기준으로 집계.
      required: [activeUsers, totalQuotaUsagePercent, companyPlan]
      properties:
        activeUsers: { type: integer, format: int64, example: 7 }
        totalQuotaUsagePercent: { type: integer, example: 72 }
        companyPlan: { type: string, enum: [FREE, STANDARD, ENTERPRISE], nullable: true, example: STANDARD }
    AdminPlanQuotaResponse:
      type: object
      required: [content, page, size, totalElements, stats]
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/AdminPlanQuotaMember" }
        page: { type: integer, example: 1 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 8 }
        stats: { $ref: "#/components/schemas/AdminPlanQuotaStats" }
    AdminPlanQuotaApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/AdminPlanQuotaResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    PlatformAdminPlanQuotaUser:
      type: object
      description: 전사 사용자 1명의 소속 회사 플랜과 이번 달 쿼터 사용 현황 한 행(#624). companyId/companyName/
        plan/quotaLimit/remainingDays 는 사용자별 소속 회사에 따라 다를 수 있다(AdminPlanQuotaMember 와 달리
        단일 값 아님). 회사 미소속(개인 계정)이면 companyId/companyName 도 null.
      required: [id, name, email, companyId, companyName, plan, quotaUsed, quotaLimit, remainingDays, status]
      properties:
        id: { type: integer, format: int64, example: 12 }
        name: { type: string, example: "김민준" }
        email: { type: string, example: "minjun.kim@company.com" }
        companyId: { type: integer, format: int64, nullable: true, example: 3 }
        companyName: { type: string, nullable: true, example: "그린타워시설관리" }
        plan: { type: string, enum: [FREE, STANDARD, ENTERPRISE], nullable: true, example: STANDARD }
        quotaUsed: { type: integer, example: 145 }
        quotaLimit: { type: integer, nullable: true, example: 1000 }
        remainingDays: { type: integer, nullable: true, example: 12 }
        status: { type: string, enum: [ACTIVE, WARNING, EXPIRED], example: ACTIVE }
    PlatformAdminPlanQuotaStats:
      type: object
      description: KPI 카드 값 — 검색어(keyword)와 무관하게 전사 기준으로 집계.
      required: [activeUsers, totalQuotaUsagePercent]
      properties:
        activeUsers: { type: integer, format: int64, example: 42 }
        totalQuotaUsagePercent: { type: integer, example: 55 }
    PlatformAdminPlanQuotaResponse:
      type: object
      required: [content, page, size, totalElements, stats]
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/PlatformAdminPlanQuotaUser" }
        page: { type: integer, example: 1 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 42 }
        stats: { $ref: "#/components/schemas/PlatformAdminPlanQuotaStats" }
    PlatformAdminPlanQuotaApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/PlatformAdminPlanQuotaResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    PlatformAdminPlanItem:
      type: object
      description: 요금제 카탈로그 항목(#624 후속). AdminPlanItem(#507)과 필드 동일 — plans 테이블이
        회사 스코프 없는 전역 데이터라 두 화면이 같은 데이터를 본다.
      required: [id, name, maxFacilities, maxMonthlyAnalyses, maxSeats, hasPdfWatermark,
        hasCounselorAccess, hasAiAddon, priceMonthly]
      properties:
        id: { type: integer, format: int64, example: 2 }
        name: { type: string, enum: [FREE, STANDARD, ENTERPRISE], example: STANDARD }
        maxFacilities: { type: integer, nullable: true, example: 10 }
        maxMonthlyAnalyses: { type: integer, nullable: true, example: 1000 }
        maxSeats: { type: integer, nullable: true, example: 5 }
        hasPdfWatermark: { type: boolean, example: false }
        hasCounselorAccess: { type: boolean, example: true }
        hasAiAddon: { type: boolean, example: true }
        priceMonthly: { type: number, nullable: true, example: 29000 }
    PlatformAdminPlanCatalogResponse:
      type: object
      required: [plans]
      properties:
        plans:
          type: array
          items: { $ref: "#/components/schemas/PlatformAdminPlanItem" }
    PlatformAdminPlanCatalogApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/PlatformAdminPlanCatalogResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    PlatformAdminPlanPolicyUpdateEntry:
      type: object
      description: PUT /api/platform-admin/plans 요청의 플랜 1건. maxFacilities/maxMonthlyAnalyses/
        maxSeats 는 null(무제한)이거나 1 이상이어야 한다. hasAiAddon 은 없음(편집 대상 아님, 기존값 유지).
      required: [name, priceMonthly, hasPdfWatermark, hasCounselorAccess]
      properties:
        name: { type: string, enum: [FREE, STANDARD, ENTERPRISE] }
        priceMonthly: { type: number, minimum: 0, example: 39000 }
        maxFacilities: { type: integer, nullable: true, minimum: 1, example: 20 }
        maxMonthlyAnalyses: { type: integer, nullable: true, minimum: 1, example: 2000 }
        maxSeats: { type: integer, nullable: true, minimum: 1, example: 10 }
        hasPdfWatermark: { type: boolean }
        hasCounselorAccess: { type: boolean }
    PlatformAdminPlanPolicyUpdateRequest:
      type: object
      description: FREE/STANDARD/ENTERPRISE 각각 정확히 한 번씩, 총 3건이어야 한다.
      required: [plans]
      properties:
        plans:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: "#/components/schemas/PlatformAdminPlanPolicyUpdateEntry" }

    # ── 대시보드 개요 집계 (구현됨, backend/core/dashboard, HAJA-17 / dev-03-01) ──
...